+++ communicator/metadata/config.json
@@ -1,7 +1,7 @@
 {
 		"mesosactionip": "10.10.174.91",
 		"mesosactionport": "5555",
-		"serveraddress": "129.236.238.130:8080",
+		"serveraddress": "192.168.1.112:8080",
 		"targettype":"Mesos",
 		"nameoraddress":  "mesos1",
 		"username":"mesos_user1",
